Put a breakpoint into PyCharm's visual debugger at the recently changed S.ComplexInfinity line inside Zero._eval_power(self, expt):     if expt.is_negative:         return S.ComplexInfinity
If I am reading my debugger correctly, it appears that we end up doing calculations like the following:
-7/4 - 13/(6*(15/8 + sqrt(1551)/18)**(1/3)) + 2*(15/8 + sqrt(1551)/18)**(1/3)

At a binary precision of 2, this becomes:
[-1.8, 3.2, -1.4]
i.e. when passed into Add(), this collapses to exactly S.Zero.

If we repeat the calculation with a higher precision we see collapsing to S.Zero is incorrect, e.g. prec=3:
[-1.75, 3.19, -1.36] == 0.08

S.Zero then feeds into taking negative powers of S.Zero, i.e. generating S.ComplexInfinity for which it cannot be determined if the value is actually real.

I am not sure if this actually fixes the underlying issue, but the recursive stack trees in this area are so complicated I honestly cannot wrap my head around them, so just opting for the later calculation on the assumption it could be a premature optimisation.